backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.0
pydantic-settings==2.5.0
pydantic_core==2.23.2
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.25
starlette==0.38.6
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.30.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
